Moving models and exceptions from google health service to correct packages
