+++ src/main/kotlin/org/radarbase/push/integration/google/exceptions/TransientGoogleHealthException.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2026 King's College London
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.radarbase.push.integration.google.exceptions
+
+/**
+ * Thrown when a retryable Google Health API error (429 or 5xx) exhausts its retry budget.
+ * Redis cursor past a chunk that never completed, the next scheduled iteration will retry.
+ */
+class TransientGoogleHealthException(message: String) : RuntimeException(message)
